Patterns for use of access control.

+++ hs_access_control/management/commands/community.py
@@ -8,12 +8,11 @@ This code is not a design pattern for actually interacting with communities.
 """
 
 from django.core.management.base import BaseCommand
-from django.contrib.auth.models import User, Group
 from hs_access_control.models.community import Community
 from hs_access_control.models.privilege import PrivilegeCodes, \
         UserGroupPrivilege, UserCommunityPrivilege, GroupCommunityPrivilege
-
-import re
+from hs_access_control.management.utilities import community_from_name_or_id, \
+        group_from_name_or_id, user_from_name
 
 
 def usage():
@@ -45,76 +44,6 @@ def usage():
     print("              [blank]: list community owners")
     print("              add: add an owner for the community.")
     print("              remove: remove an owner from the community.")
-
-
-def group_from_name_or_id(gname):
-    """ return a group object given either an id or a name """
-
-    RE_INT = re.compile(r'^([1-9]\d*|0)$')
-    if RE_INT.match(gname):
-        try:
-            gid = int(gname)
-            group = Group.objects.get(id=gid)
-            return group
-        except Group.DoesNotExist:
-            print("group with id {} does not exist.".format(str(gid)))
-            usage()
-            exit(1)
-
-    else:  # interpret as name
-        groups = Group.objects.filter(name=gname)
-        if groups.count() == 0:
-            print("group '{}' not found.".format(gname))
-            usage()
-            exit(1)
-        elif groups.count() == 1:
-            group = groups[0]
-            return group
-        else:
-            print("Group name {} is not unique. Please use group id instead:".format(gname))
-            for g in groups:
-                print("   '{}' (id={})".format(g.name, str(g.id)))
-            exit(1)
-
-
-def community_from_name_or_id(cname):
-    """ return a group object given either an id or a name """
-
-    RE_INT = re.compile(r'^([1-9]\d*|0)$')
-    if RE_INT.match(cname):
-        try:
-            cid = int(cname)
-            group = Community.objects.get(id=cid)
-            return group
-        except Community.DoesNotExist:
-            print("community with id {} does not exist.".format(str(cid)))
-            usage()
-            exit(1)
-
-    else:  # interpret as name
-        communities = Community.objects.filter(name=cname)
-        if communities.count() == 0:
-            print("community with name '{}' does not exist.".format(cname))
-            usage()
-            exit(1)
-        elif communities.count() == 1:
-            community = communities[0]
-            return community
-        else:
-            print("Community name '{}' is not unique. Please use community id instead:"
-                  .format(cname))
-            for g in communities:
-                print("   '{}' (id={})".format(g.name, str(g.id)))
-            exit(1)
-
-
-def user_from_name(uname):
-    try:
-        return User.objects.get(username=uname)
-    except User.DoesNotExist:
-        print("user with username '{}' does not exist.".format(uname))
-        usage()
-        exit(1)
 
 
 class Command(BaseCommand):
@@ -175,7 +104,11 @@ class Command(BaseCommand):
             oname = options['owner']
         else:
             oname = 'admin'
+
         owner = user_from_name(oname)
+        if owner is None:
+            usage()
+            exit(1)
 
         if options['allow_edit']:  # this is a group privilege
             privilege = PrivilegeCodes.CHANGE
@@ -191,6 +124,10 @@ class Command(BaseCommand):
 
         if command is None or command == 'list':
             community = community_from_name_or_id(cname)
+            if community is None:
+                usage()
+                exit(1)
+
             print("community '{}' (id={}):".format(community.name, community.id))
             print("  description: {}".format(community.description))
             print("  purpose: {}".format(community.purpose))
@@ -250,6 +187,9 @@ class Command(BaseCommand):
         elif command == 'owner':
             # at this point, community must exist
             community = community_from_name_or_id(cname)
+            if community is None:
+                usage()
+                exit(1)
 
             if len(options['command']) < 3:
                 # list owners
@@ -261,6 +201,10 @@ class Command(BaseCommand):
 
             oname = options['command'][2]
             owner = user_from_name(oname)
+            if owner is None:
+                usage()
+                exit(1)
+
             if len(options['command']) < 4:
                 print("user {} owns community '{}' (id={})"
                       .format(owner.username, community.name, str(community.id)))
@@ -283,8 +227,12 @@ class Command(BaseCommand):
                 exit(1)
 
         elif command == 'group':
+
             # at this point, community must exist
             community = community_from_name_or_id(cname)
+            if community is None:
+                usage()
+                exit(1)
 
             # not specifying a group should list groups
             if len(options['command']) < 3:
@@ -304,6 +252,9 @@ class Command(BaseCommand):
 
             gname = options['command'][2]
             group = group_from_name_or_id(gname)
+            if group is None:
+                usage()
+                exit(1)
 
             if len(options['command']) < 4:
                 print("community groups: no action specified.")
@@ -358,6 +309,10 @@ class Command(BaseCommand):
 
         elif command == 'remove':
             community = community_from_name_or_id(cname)
+            if community is None:
+                usage()
+                exit(1)
+
             print("removing community '{}' (id={}).".format(community.name, community.id))
             community.delete()
